bootloader_message: support for U-Boot environment
As in previous Digi Embedded Android releases, we use the U-Boot environment
instead of a /misc partition to store the bootloader_message.

This commit adds supports to write/read to/from used both during factory reset
and firmware update. For that it uses the prebuilt fw_printenv binary included
in device/digi repository.

https://jira.digi.com/browse/ADK4A-1326

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>