recovery: add support to set the CAAM protected fs encryption key
This new command uses the TrustFence tool to set the file system
encryption key.

The encryption key can be set directly from u-boot:

  => env set recovery_command 'recovery --encryption_key=[KEY]'
  => env save
  => run recoverycmd

Where [KEY] is a base64 encoded 32-bytes (256 bits) key. The KEY argument
is optional, if not provided a random encryption key will be created.

Note that at the moment it is not possible to change the encryption key
and maintain the data, so every time the encryption key is changed the
data partition will be wiped.

This commit is a forward port from code in Lollipop, with some rework
and simplifications.

https://jira.digi.com/browse/ADK4A-1773

Co-authored-by: Javier Viguera <javier.viguera@digi.com>

Signed-off-by: David Escalona <david.escalona@digi.com>
Signed-off-by: Javier Viguera <javier.viguera@digi.com>